Refactor routes to avoid heavy single module
Routes are split by those that just access journal
information and do not need data files and those
that access NeXus information from the run files.
